Fixed eventlog and added argument autocomplete
