+++ src/condor_tools/typer_htcondor/job.py
@@ -1,12 +1,11 @@
 # job.py
 import typer
+job_app = typer.Typer()
 
-@job_app.command()(status)
+@job_app.command()
 def status():
     typer.echo("Checking job status...")
 
-@job_app.command()(submit)
+@job_app.command()
 def submit():
     typer.echo("Submitting job...")
-
-job_app = typer.Typer()
